Refactor NetworkModule
